tokio: encapsulate taskdumping refcount shenanigans
ref: https://github.com/tokio-rs/tokio/pull/5608/files#r1161100780
