+++ tokio/src/runtime/scheduler/current_thread.rs
@@ -380,9 +380,10 @@ impl Handle {
     /// Capture a snapshot of this runtime's state.
     #[cfg(all(tokio_unstable, feature = "taskdump"))]
     pub(crate) fn dump(&self) -> crate::runtime::Dump {
-        use crate::runtime::{dump, task::trace::Trace};
+        use crate::runtime::dump;
+        use task::trace::trace_current_thread;
 
-        let mut snapshots = vec![];
+        let mut traces = vec![];
 
         // todo: how to make this work outside of a runtime context?
         CURRENT.with(|maybe_context| {
@@ -390,39 +391,30 @@ impl Handle {
             let context = if let Some(context) = maybe_context {
                 context
             } else {
-                return
+                return;
             };
             let mut maybe_core = context.core.borrow_mut();
             let core = if let Some(core) = maybe_core.as_mut() {
                 core
             } else {
-                return
+                return;
             };
             let local = &mut core.tasks;
-            let _ = local.drain(..);
 
-            // drain the injection queue
-            if let Some(injection) = self.shared.queue.lock().as_mut() {
-                let _ = injection.drain(..);
-            }
+            let mut injection = self.shared.queue.lock();
+            let injection = if let Some(injection) = injection.as_mut() {
+                injection
+            } else {
+                return;
+            };
 
-            // notify each task
-            let mut tasks = vec![];
-            self.shared.owned.for_each(|task| {
-                // set the notified bit
-                let _ = task.as_raw().state().transition_to_notified_for_tracing();
-                // store the raw tasks into a vec
-                tasks.push(task.as_raw());
-            });
-
-            // trace each task
-            for task in tasks {
-                let ((), trace) = Trace::capture(|| task.poll());
-                snapshots.push(dump::Task::new(trace));
-            }
+            traces = trace_current_thread(&self.shared.owned, local, injection)
+                .into_iter()
+                .map(dump::Task::new)
+                .collect();
         });
 
-        dump::Dump::new(snapshots)
+        dump::Dump::new(traces)
     }
 
     fn pop(&self) -> Option<task::Notified<Arc<Handle>>> {
